feat(scheduler): add per-model dispatch with priority scoring
- Per-model dispatch queues with semaphore bounded by max_inflight
- ModelDeadlineScheduler: per-model heaps with deadline-aware scoring
- Priority travels with every observation (field 8 in proto); 0 = highest
- Python, C++, Rust clients all serialize priority
- Fix per-model response routing on shared connections (client-side demux)
- Default scheduling strategy updated to model_deadline
- Four bug fixes: Rust ETERM panic, default strategy, batch_by_id collision, BFLOAT16 mapping
- Update all examples, docs, and tests

+++ python/inferential/proto/__init__.py
@@ -67,7 +67,7 @@ _DTYPE_TO_NUMPY: dict[DType, np.dtype] = {
     FLOAT16: np.dtype("float16"),
     FLOAT32: np.dtype("float32"),
     FLOAT64: np.dtype("float64"),
-    BFLOAT16: np.dtype("float16"),  # closest numpy equivalent
+    BFLOAT16: np.dtype("uint16"),  # bfloat16 has no numpy equivalent; raw bytes preserved
     UINT8: np.dtype("uint8"),
     INT32: np.dtype("int32"),
     INT64: np.dtype("int64"),
